Module Prelude.List

include module type of struct include Stdlib.List end

type !'a t = 'a list =

| []

| :: of 'a * 'a list

val length : 'a list -> int

val compare_lengths : 'a list -> 'b list -> int

val compare_length_with : 'a list -> int -> int

val cons : 'a -> 'a list -> 'a list

val singleton : 'a -> 'a list

val hd : 'a list -> 'a

val tl : 'a list -> 'a list

val nth : 'a list -> int -> 'a

val nth_opt : 'a list -> int -> 'a option

val rev : 'a list -> 'a list

val init : int -> (int -> 'a) -> 'a list

val append : 'a list -> 'a list -> 'a list

val rev_append : 'a list -> 'a list -> 'a list

val concat : 'a list list -> 'a list

val flatten : 'a list list -> 'a list

val equal : ('a -> 'a -> bool) -> 'a list -> 'a list -> bool

val compare : ('a -> 'a -> int) -> 'a list -> 'a list -> int

val iter : ('a -> unit) -> 'a list -> unit

val iteri : (int -> 'a -> unit) -> 'a list -> unit

val map : ('a -> 'b) -> 'a list -> 'b list

val mapi : (int -> 'a -> 'b) -> 'a list -> 'b list

val rev_map : ('a -> 'b) -> 'a list -> 'b list

val filter_map : ('a -> 'b option) -> 'a list -> 'b list

val concat_map : ('a -> 'b list) -> 'a list -> 'b list

val fold_left_map : ('acc -> 'a -> 'acc * 'b) -> 'acc -> 'a list -> 'acc * 'b list

val fold_left : ('acc -> 'a -> 'acc) -> 'acc -> 'a list -> 'acc

val fold_right : ('a -> 'acc -> 'acc) -> 'a list -> 'acc -> 'acc

val iter2 : ('a -> 'b -> unit) -> 'a list -> 'b list -> unit

val map2 : ('a -> 'b -> 'c) -> 'a list -> 'b list -> 'c list

val rev_map2 : ('a -> 'b -> 'c) -> 'a list -> 'b list -> 'c list

val fold_left2 : ('acc -> 'a -> 'b -> 'acc) -> 'acc -> 'a list -> 'b list -> 'acc

val fold_right2 : ('a -> 'b -> 'acc -> 'acc) -> 'a list -> 'b list -> 'acc -> 'acc

val for_all : ('a -> bool) -> 'a list -> bool

val exists : ('a -> bool) -> 'a list -> bool

val for_all2 : ('a -> 'b -> bool) -> 'a list -> 'b list -> bool

val exists2 : ('a -> 'b -> bool) -> 'a list -> 'b list -> bool

val mem : 'a -> 'a list -> bool

val memq : 'a -> 'a list -> bool

val find : ('a -> bool) -> 'a list -> 'a

val find_opt : ('a -> bool) -> 'a list -> 'a option

val find_index : ('a -> bool) -> 'a list -> int option

val find_map : ('a -> 'b option) -> 'a list -> 'b option

val find_mapi : (int -> 'a -> 'b option) -> 'a list -> 'b option

val filter : ('a -> bool) -> 'a list -> 'a list

val find_all : ('a -> bool) -> 'a list -> 'a list

val filteri : (int -> 'a -> bool) -> 'a list -> 'a list

val take : int -> 'a list -> 'a list

val drop : int -> 'a list -> 'a list

val take_while : ('a -> bool) -> 'a list -> 'a list

val drop_while : ('a -> bool) -> 'a list -> 'a list

val partition : ('a -> bool) -> 'a list -> 'a list * 'a list

val partition_map : ('a -> ('b, 'c) Stdlib.Either.t) -> 'a list -> 'b list * 'c list

val assoc : 'a -> ('a * 'b) list -> 'b

val assoc_opt : 'a -> ('a * 'b) list -> 'b option

val assq : 'a -> ('a * 'b) list -> 'b

val assq_opt : 'a -> ('a * 'b) list -> 'b option

val mem_assoc : 'a -> ('a * 'b) list -> bool

val mem_assq : 'a -> ('a * 'b) list -> bool

val remove_assoc : 'a -> ('a * 'b) list -> ('a * 'b) list

val remove_assq : 'a -> ('a * 'b) list -> ('a * 'b) list

val split : ('a * 'b) list -> 'a list * 'b list

val combine : 'a list -> 'b list -> ('a * 'b) list

val sort : ('a -> 'a -> int) -> 'a list -> 'a list

val stable_sort : ('a -> 'a -> int) -> 'a list -> 'a list

val fast_sort : ('a -> 'a -> int) -> 'a list -> 'a list

val sort_uniq : ('a -> 'a -> int) -> 'a list -> 'a list

val merge : ('a -> 'a -> int) -> 'a list -> 'a list -> 'a list

val to_seq : 'a list -> 'a Stdlib.Seq.t

val of_seq : 'a Stdlib.Seq.t -> 'a list

val return : 'a -> 'a t

val is_empty : 'a t -> bool

val is_one : 'a t -> bool

val first : 'a t -> 'a option

val first_partition : 'a t -> 'a option * 'a t

val last_partition : 'a t -> 'a option * 'a t

val last : 'a t -> 'a option

val ints : int -> int list

val flatten_opt : 'a option list -> 'a t * bool

val zip3 : 'a t -> 'b t -> 'c t -> ('a * 'b * 'c) t

val zip4 : 'a t -> 'b t -> 'c t -> 'd t -> ('a * 'b * 'c * 'd) t

val split3 : ('a * 'b * 'c) t -> 'a t * 'b t * 'c t

val split4 : ('a * 'b * 'c * 'd) t -> 'a t * 'b t * 'c t * 'd t

val cartesian : 'a list -> 'b list -> ('a * 'b) list

Cartesian product of 2 lists.

val map_cartesian : ('a -> 'b -> 'c) -> 'a list -> 'b list -> 'c list

val flat_cartesian : 'a list list -> 'a list list -> 'a list list

Makes cartesian product of two lists of lists as another list of lists.

val powerset : 'a t -> 'a t t

Returns a powerset (all subsets) of the given list.

val powerset_nz : 'a t -> 'a t list

Returns powerset without the empty list.

val powerset_partition : 'a t -> ('a t * 'a t) t

val flat_map : ('a -> 'b list) -> 'a list -> 'b list

val general_cartesian : 'a list list -> 'a t t

val unfold_for : ('b -> ('a * 'b) option) -> 'b -> int -> 'a list

val unfold_while : ('b -> ('a * 'b) option) -> 'b -> ('a -> bool) -> 'a list

val unfold_for_while : ('b -> ('a * 'b) option) -> 'b -> int -> ('a -> bool) -> 'a list * bool

val drop_nth : 'a t -> int -> 'a t

val any : bool list -> bool

val all : bool list -> bool

val flat : 'a option list -> 'a list

val minmax : ('a -> 'a -> int) -> 'a t -> 'a * 'a

val find_opt_partition : ('a -> bool) -> 'a t -> 'a option * 'a t

val print : ('a -> unit) -> 'a list -> unit

val argmax_opt : ('a -> 'a -> int) -> 'a t -> int option

val to_string : ?sep:string -> ('a -> string) -> 'a list -> string

val map_inplace_or_drop : ('a -> [< `Drop | `Replace of 'a | `Skip ]) -> 'a t -> 'a option * 'a t

Replaces an element inplace with `Replace, drops when `Drop and skips if `Skip

val map_inplace_once_indicate : ('a -> 'a option) -> 'a t -> 'a t * bool

Modifies first element encountered when f x = Some _, returns _, true if happened.

val map_inplace_once : ('a -> 'a option) -> 'a t -> 'a t

Modifies first element encountered when f x = Some _

val map_inplace_once_or : ('a -> 'a option) -> ('a t -> 'a t) -> 'a t -> 'a t

val filter_mapi : (int -> 'a -> 'b option) -> 'a t -> 'b t

val reduce_left : ('a -> 'b -> 'a) -> ('b -> 'a) -> 'b t -> 'a

reduce_left is similar to fold_left, except it uses acc l0 as its accumulator.

val reduce_right : ('a -> 'b -> 'b) -> ('a -> 'b) -> 'a t -> 'b

reduce_right is similar to fold_right, except is uses acc ltail as its accumulator.

val zip_list : 'a list list -> 'a list list

zip_list zips list of lists from list of sequences. Ends when any outer list ends.

val pair_apply : ('a -> 'a -> 'a) -> 'a t -> 'a t

val fold_merge : ('a -> 'a -> 'a) -> 'a t -> 'a

val uncons : 'a t -> 'a * 'a list

val fold_lefti : ('a -> int -> 'b -> 'a) -> 'a -> 'b t -> 'a

val fold_left_mapr : ('a -> 'b -> ('a * 'c, 'd) Stdlib.result) -> 'a -> 'b t -> ('a * 'c list, 'd) Stdlib.result

val fold_leftr : ('a -> 'b -> ('a, 'c) Stdlib.result) -> 'a -> 'b t -> ('a, 'c) Stdlib.result

val insert : 'a t -> 'a -> int -> 'a t