Module Backend.Machine

val diff_counter_update : 'a -> ('b, 'c) Symbolic__Machine.int_expr -> ('b, 'c) Symbolic__Machine.bool_expr -> ('b, 'c) Symbolic__Machine.bool_expr -> 'a * ('b, 'c) Symbolic__Machine.expr

val clocks_to_bvars : 'a list -> ('b, 'a) Symbolic.Machine.bool_expr list

val build_excl_dec_tree : ('a, 'b) Symbolic__Machine.int_expr option -> ('a, 'b) Symbolic.Machine.bool_expr list -> ('a, 'b) Symbolic__Machine.bool_expr

Builds a decision tree where only one variable can be true at the time, and has to be related to a specific choice_var value.

val iparam_to_string : (string, int) Mrtccsl__.Language.Cstr.arg -> string

val rparam_to_string : (string, Mpq.f Number.Rational.tt) Mrtccsl__.Language.Cstr.arg -> string

val iparam_to_expr : (Symbolic.Machine.Interpretation.var, int) Mrtccsl__.Language.Cstr.arg -> ('a, Symbolic.Machine.Interpretation.var) Symbolic__Machine.int_expr

val rparam_to_expr : (Symbolic.Machine.Interpretation.var, Number.Rational.t) Mrtccsl__.Language.Cstr.arg -> ('a, Symbolic.Machine.Interpretation.var) Symbolic__Machine.rat_expr

Stateless constraints

val stateless_as_machine : ('a, 'b) Symbolic__Machine.bool_expr -> ('a, 'b) Symbolic__Machine.t

val stateless_with_range_cond : int -> ('a, 'b) Symbolic__Machine.bool_expr -> ('a, 'b) Symbolic__Machine.int_expr Prelude.Option.t -> ('a, 'b) Symbolic__Machine.t

Helper function that adds a range condition on integer varaible (if present) and returns stateless machine.

val exclusion_as_machine : 'a list -> 'a option -> ('b, 'a) Symbolic__Machine.t

val coincidence_as_machine : 'a list -> ('b, 'a) Symbolic__Machine.t

val subclocking_as_machine : 'a -> 'a -> 'a option -> ('b, 'a) Symbolic__Machine.t

val minus_as_machine : 'a -> 'a -> 'a list -> ('b, 'a) Symbolic__Machine.t

val union_as_machine : 'a -> 'a Prelude.List.t -> ('b, 'a) Symbolic__Machine.t

val disj_union_as_machine : 'a -> 'a Prelude.List.t -> 'a option -> ('b, 'a) Symbolic__Machine.t

val intersection_as_machine : 'a -> 'a Prelude.List.t -> ('b, 'a) Symbolic__Machine.t

Stateful constraints

val causality_as_machine : strict:bool -> string -> string -> (string, string) Symbolic__Machine.t

val delay_vars : 'a -> string -> (Symbolic.Machine.Interpretation.var, int) Mrtccsl__.Language.Cstr.arg -> string -> string * (string, 'b) Symbolic.Machine.int_queue_expr * ('c, 'a) Symbolic.Machine.bool_expr * ('d, string) Symbolic.Machine.bool_expr * ('e, Symbolic.Machine.Interpretation.var) Symbolic__Machine.int_expr * ('f, string) Symbolic.Machine.bool_expr

val delay_as_machine : Symbolic.Machine.Interpretation.var -> Symbolic.Machine.Interpretation.var -> (Symbolic.Machine.Interpretation.var, int) Mrtccsl__.Language.Cstr.arg -> Symbolic.Machine.Interpretation.var -> (string, Symbolic.Machine.Interpretation.var) Symbolic__Machine.t

Implements logical delay as abstract machine. Materializes delay at the moment when arg ticks.

val delay_as_late_acceptor : Symbolic.Machine.Interpretation.var -> Symbolic.Machine.Interpretation.var -> (Symbolic.Machine.Interpretation.var, int) Mrtccsl__.Language.Cstr.arg -> Symbolic.Machine.Interpretation.var -> (string, Symbolic.Machine.Interpretation.var) Symbolic__Machine.t

Implements logical delay as abstract machine. Important difference with delay_as_machine: checks correctness of delay between out and arg (in terms of base) at the moment when out ticks.

val alternate_as_machine : string -> string -> bool -> (string, string) Symbolic__Machine.t

val sample_as_machine : string -> string -> string -> (string, string) Symbolic__Machine.t

val slowest_fastest_as_machine : slowest:bool -> string -> string list -> (string, string) Symbolic__Machine.t

val periodic_as_machine : Symbolic.Machine.Interpretation.var -> Symbolic.Machine.Interpretation.var -> int -> (Symbolic.Machine.Interpretation.var, int) Mrtccsl__.Language.Cstr.arg -> (Symbolic.Machine.Interpretation.var, int) Mrtccsl__.Language.Cstr.arg -> (string, Symbolic.Machine.Interpretation.var) Symbolic__Machine.t

val periodic_as_late_acceptor : Symbolic.Machine.Interpretation.var -> Symbolic.Machine.Interpretation.var -> int -> (Symbolic.Machine.Interpretation.var, int) Mrtccsl__.Language.Cstr.arg -> (Symbolic.Machine.Interpretation.var, int) Mrtccsl__.Language.Cstr.arg -> (string, Symbolic.Machine.Interpretation.var) Symbolic__Machine.t

val first_sampled_as_machine : string -> string -> string -> (string, string) Symbolic__Machine.t

val last_sampled_as_machine : string -> string -> string -> (string, string) Symbolic__Machine.t

val forbid_as_machine : string -> string -> string list -> (string, string) Symbolic__Machine.t

val allow_as_machine : string -> string -> string list -> (string, string) Symbolic__Machine.t

val mutex_as_machine : (string * string) list -> (string, string) Symbolic__Machine.t

val rtdelay_vars : 'a -> string -> (Symbolic.Machine.Interpretation.var, Mpq.f Number.Rational.tt) Mrtccsl__.Language.Cstr.arg -> string * (string, 'b) Symbolic__Machine.rat_queue_expr * ('c, 'a) Symbolic__Machine.bool_expr * ('d, string) Symbolic__Machine.bool_expr * ('e, Symbolic.Machine.Interpretation.var) Symbolic__Machine.rat_expr

val rtdelay_as_machine : now:(string, Symbolic.Machine.Interpretation.var) Symbolic__Machine.rat_expr -> Symbolic.Machine.Interpretation.var -> Symbolic.Machine.Interpretation.var -> (Symbolic.Machine.Interpretation.var, Mpq.f Number.Rational.tt) Mrtccsl__.Language.Cstr.arg -> (string, Symbolic.Machine.Interpretation.var) Symbolic__Machine.t

val rtdelay_as_late_acceptor : now:(string, Symbolic.Machine.Interpretation.var) Symbolic__Machine.rat_expr -> Symbolic.Machine.Interpretation.var -> Symbolic.Machine.Interpretation.var -> (Symbolic.Machine.Interpretation.var, Mpq.f Number.Rational.tt) Mrtccsl__.Language.Cstr.arg -> (string, Symbolic.Machine.Interpretation.var) Symbolic__Machine.t

val rtperiodic_vars : string -> Number.Rational.t -> (Symbolic.Machine.Interpretation.var, Number.Rational.t) Mrtccsl__.Language.Cstr.arg -> (Symbolic.Machine.Interpretation.var, Number.Rational.t) Mrtccsl__.Language.Cstr.arg -> string * (string, 'a) Symbolic__Machine.rat_expr * ('b, string) Symbolic__Machine.bool_expr * ('c, 'd) Symbolic__Machine.rat_expr * ('e, Symbolic.Machine.Interpretation.var) Symbolic__Machine.rat_expr * ('f, Symbolic.Machine.Interpretation.var) Symbolic__Machine.rat_expr

val drift_periodic_as_machine : now:(string, Symbolic.Machine.Interpretation.var) Symbolic__Machine.rat_expr -> Symbolic.Machine.Interpretation.var -> Number.Rational.t -> (Symbolic.Machine.Interpretation.var, Number.Rational.t) Mrtccsl__.Language.Cstr.arg -> (Symbolic.Machine.Interpretation.var, Number.Rational.t) Mrtccsl__.Language.Cstr.arg -> (string, Symbolic.Machine.Interpretation.var) Symbolic__Machine.t

val jitter_periodic_as_machine : now:(string, Symbolic.Machine.Interpretation.var) Symbolic__Machine.rat_expr -> Symbolic.Machine.Interpretation.var -> Number.Rational.t -> (Symbolic.Machine.Interpretation.var, Number.Rational.t) Mrtccsl__.Language.Cstr.arg -> (Symbolic.Machine.Interpretation.var, Number.Rational.t) Mrtccsl__.Language.Cstr.arg -> (string, Symbolic.Machine.Interpretation.var) Symbolic__Machine.t

val sporadic_as_machine : now:(string, Symbolic.Machine.Interpretation.var) Symbolic__Machine.rat_expr -> Symbolic.Machine.Interpretation.var -> (Symbolic.Machine.Interpretation.var, Number.Rational.t) Mrtccsl__.Language.Cstr.arg -> bool -> (string, Symbolic.Machine.Interpretation.var) Symbolic__Machine.t

val constraint_as_machine : Symbolic.Machine.Interpretation.var -> (Symbolic.Machine.Interpretation.var, Symbolic.Machine.Interpretation.var, Symbolic.Machine.Interpretation.var, Symbolic.Machine.Interpretation.var, Symbolic.Machine.Interpretation.var, Mpq.f Number.Rational.tt) Mrtccsl__.Language.Cstr.clock_constr -> (string, string) Symbolic.Machine.t

Converts constraint into an abstract machine.

val empty : string * (string, string) Symbolic.Machine.t

Empty (as it does not constrain any clocks) machine with the basic condition of strict monotonicity on the real-time progression.

val numerical_relation_as_machine : ('a -> 'b) -> (('a, 'c) Mrtccsl__.Language.Cstr.arg -> 'd) -> (('b * Expr.num_rel * 'd) -> ('e, 'f) Symbolic__Machine.bool_expr) -> ('a, 'c) Mrtccsl__.Language.Cstr.numeric_rel_constr -> ('e, 'f) Symbolic__Machine.t

Converts numerical relation into an abstract machine.

type sim = Symbolic.Machine.Interpretation.var * (Symbolic.Machine.Interpretation.var, Symbolic.Machine.Interpretation.var) Symbolic.Machine.t

val of_spec : ?debug:'a -> (Symbolic.Machine.Interpretation.var, Symbolic.Machine.Interpretation.var, Symbolic.Machine.Interpretation.var, Symbolic.Machine.Interpretation.var, Symbolic.Machine.Interpretation.var, Mpq.f Number.Rational.tt) Mrtccsl__.Language.Specification.t -> sim

Converts the specification constraints into a synchronized abstract machine.

val step_as_inputs : Symbolic.Machine.VarMap.key -> (Symbolic.Machine.VarMap.key list, Number.Rational.t) Trace.step -> Symbolic.Machine.Interpretation.inputs

val sexp_of_step : (string list, Mpq.f Number.Rational.tt) Trace.step -> Sexplib0.Sexp.t

val accept_trace : (Symbolic.Machine.VarMap.key * (Prelude.String.t, Symbolic.Machine.Interpretation.var) Symbolic__Machine.t) -> (Symbolic.Machine.VarMap.key list, Number.Rational.t) Trace.step Prelude.Seq.t -> bool

Checks if machine accepts a trace.