Module Language.Cstr

Constraint definitions and supporing functions.

type ('v, 'n) arg =

| Var of 'v

| Const of 'n

Type of constraint arguments.

val map_arg : ('a -> 'b) -> ('c -> 'd) -> ('a, 'c) arg -> ('b, 'd) arg

val pp_arg : 'v 'n. (Ppx_deriving_runtime.Format.formatter -> 'v -> Ppx_deriving_runtime.unit) -> (Ppx_deriving_runtime.Format.formatter -> 'n -> Ppx_deriving_runtime.unit) -> Ppx_deriving_runtime.Format.formatter -> ('v, 'n) arg -> Ppx_deriving_runtime.unit

val show_arg : 'v 'n. (Ppx_deriving_runtime.Format.formatter -> 'v -> Ppx_deriving_runtime.unit) -> (Ppx_deriving_runtime.Format.formatter -> 'n -> Ppx_deriving_runtime.unit) -> ('v, 'n) arg -> Ppx_deriving_runtime.string

val fold_arg : ('a -> 'b -> 'c) -> ('a -> 'd -> 'c) -> 'a -> ('b, 'd) arg -> 'c

val var : 'a -> ('a, 'b) arg

Constructs variable/parameter argument.

val const : 'a -> ('b, 'a) arg

Constructs constant argument.

val unwrap_arg : 'a 'b 'c. var:('a -> 'b) -> const:('c -> 'b) -> ('a, 'c) arg -> 'b

Rewrites the argument by using var and const functions.

type ('c, 'tp, 'ip, 'tv, 'iv, 't) clock_constr =

| Precedence of {

cause : 'c;

conseq : 'c;

}

| Causality of {

cause : 'c;

conseq : 'c;

}

| Exclusion of {

args : 'c list;

choice : 'iv option;

}

| Coincidence of 'c list

| Subclocking of {

sub : 'c;

super : 'c;

choice : 'iv option;

}

| Minus of {

out : 'c;

arg : 'c;

except : 'c list;

}

| Delay of {

out : 'c;

arg : 'c;

delay : ('iv, int) arg;

base : 'c;

}

| Fastest of {

out : 'c;

args : 'c list;

}

| Slowest of {

out : 'c;

args : 'c list;

}

| Intersection of {

out : 'c;

args : 'c list;

}

| Union of {

out : 'c;

args : 'c list;

}

| Periodic of {

out : 'c;

base : 'c;

period : int;

error : ('iv, int) arg;

offset : ('ip, int) arg;

}

| Sample of {

out : 'c;

arg : 'c;

base : 'c;

}

| Alternate of {

first : 'c;

second : 'c;

strict : bool;

}

| RTdelay of {

out : 'c;

arg : 'c;

delay : ('tv, 't) arg;

}

| CumulPeriodic of {

out : 'c;

period : 't;

error : ('tv, 't) arg;

offset : ('tp, 't) arg;

}

| AbsPeriodic of {

out : 'c;

period : 't;

error : ('tv, 't) arg;

offset : ('tp, 't) arg;

}

| FirstSampled of {

out : 'c;

arg : 'c;

base : 'c;

}

| LastSampled of {

out : 'c;

arg : 'c;

base : 'c;

}

| Forbid of {

left : 'c;

right : 'c;

args : 'c list;

}

| Allow of {

left : 'c;

right : 'c;

args : 'c list;

}

| Sporadic of {

out : 'c;

at_least : ('tp, 't) arg;

strict : bool;

}

| Pool of int * ('c * 'c) list
(*
Mutex is a special case of Pool where n=1
*)

| DisjunctiveUnion of {

out : 'c;

args : 'c list;

choice : 'iv option;

}

Clock constraint type. 'c is clock variable symbols type, '_p is parameter (unchanged during execution) symbols type ('tp is for durations, 'ip is for integers), '_v is symbol type for variables (change during execution), 't is the type of duration constants.

val map_clock_constr : ('a -> 'b) -> ('c -> 'd) -> ('e -> 'f) -> ('g -> 'h) -> ('i -> 'j) -> ('k -> 'l) -> ('a, 'c, 'e, 'g, 'i, 'k) clock_constr -> ('b, 'd, 'f, 'h, 'j, 'l) clock_constr

val pp_clock_constr : 'c 'tp 'ip 'tv 'iv 't. (Ppx_deriving_runtime.Format.formatter -> 'c -> Ppx_deriving_runtime.unit) -> (Ppx_deriving_runtime.Format.formatter -> 'tp -> Ppx_deriving_runtime.unit) -> (Ppx_deriving_runtime.Format.formatter -> 'ip -> Ppx_deriving_runtime.unit) -> (Ppx_deriving_runtime.Format.formatter -> 'tv -> Ppx_deriving_runtime.unit) -> (Ppx_deriving_runtime.Format.formatter -> 'iv -> Ppx_deriving_runtime.unit) -> (Ppx_deriving_runtime.Format.formatter -> 't -> Ppx_deriving_runtime.unit) -> Ppx_deriving_runtime.Format.formatter -> ('c, 'tp, 'ip, 'tv, 'iv, 't) clock_constr -> Ppx_deriving_runtime.unit

val show_clock_constr : 'c 'tp 'ip 'tv 'iv 't. (Ppx_deriving_runtime.Format.formatter -> 'c -> Ppx_deriving_runtime.unit) -> (Ppx_deriving_runtime.Format.formatter -> 'tp -> Ppx_deriving_runtime.unit) -> (Ppx_deriving_runtime.Format.formatter -> 'ip -> Ppx_deriving_runtime.unit) -> (Ppx_deriving_runtime.Format.formatter -> 'tv -> Ppx_deriving_runtime.unit) -> (Ppx_deriving_runtime.Format.formatter -> 'iv -> Ppx_deriving_runtime.unit) -> (Ppx_deriving_runtime.Format.formatter -> 't -> Ppx_deriving_runtime.unit) -> ('c, 'tp, 'ip, 'tv, 'iv, 't) clock_constr -> Ppx_deriving_runtime.string

val fold_clock_constr : ('a -> 'b -> 'a) -> ('a -> 'c -> 'a) -> ('a -> 'd -> 'a) -> ('a -> 'e -> 'a) -> ('a -> 'f -> 'a) -> ('a -> 'g -> 'a) -> 'a -> ('b, 'c, 'd, 'e, 'f, 'g) clock_constr -> 'a

val to_string : ('a -> string) -> ('b -> string) -> ('c -> string) -> ('d -> string) -> ('e -> string) -> ('f -> string) -> ('a, 'b, 'c, 'd, 'e, 'f) clock_constr -> string

val argument_assumptions : ('a, 'b, 'c, 'd, 'e, 'f) clock_constr -> bool

Defines constraint argument assumptions. Currently only checks for wrong usage of logical clocks.

type ('v, 'c) numeric_rel_constr =

| NumRelation of 'v * Expr.num_rel * ('v, 'c) arg

Numerical relation constraint, represents 'v |><| ('v | 'c) relation, where 'v is variable type and 'c is constant type.

val map_numeric_rel_constr : ('a -> 'b) -> ('c -> 'd) -> ('a, 'c) numeric_rel_constr -> ('b, 'd) numeric_rel_constr

val pp_numeric_rel_constr : 'v 'c. (Ppx_deriving_runtime.Format.formatter -> 'v -> Ppx_deriving_runtime.unit) -> (Ppx_deriving_runtime.Format.formatter -> 'c -> Ppx_deriving_runtime.unit) -> Ppx_deriving_runtime.Format.formatter -> ('v, 'c) numeric_rel_constr -> Ppx_deriving_runtime.unit

val show_numeric_rel_constr : 'v 'c. (Ppx_deriving_runtime.Format.formatter -> 'v -> Ppx_deriving_runtime.unit) -> (Ppx_deriving_runtime.Format.formatter -> 'c -> Ppx_deriving_runtime.unit) -> ('v, 'c) numeric_rel_constr -> Ppx_deriving_runtime.string

val fold_numeric_rel_constr : ('a -> 'b -> 'a) -> ('a -> 'c -> 'a) -> 'a -> ('b, 'c) numeric_rel_constr -> 'a

type 'n distribution =

| Uniform

| Normal of {

mean : 'n;

deviation : 'n;

}

| Exponential of {

rate : 'n;

}

Variants of continuous-valued distributions that can be used in a simulation.

val map_distribution : ('a -> 'b) -> 'a distribution -> 'b distribution

val pp_distribution : 'n. (Ppx_deriving_runtime.Format.formatter -> 'n -> Ppx_deriving_runtime.unit) -> Ppx_deriving_runtime.Format.formatter -> 'n distribution -> Ppx_deriving_runtime.unit

val show_distribution : 'n. (Ppx_deriving_runtime.Format.formatter -> 'n -> Ppx_deriving_runtime.unit) -> 'n distribution -> Ppx_deriving_runtime.string

val fold_distribution : ('a -> 'b -> 'a) -> 'a -> 'b distribution -> 'a

type ('iv, 'tv, 't) probability_constr =

| DiscreteValued of {

name : 'iv;

ratios : (int * int) list;

}

| ContinuousValued of {

name : 'tv;

dist : 't distribution;

}

Constraint that binds a distribution to variable 'iv or 'tv.

val map_probability_constr : ('a -> 'b) -> ('c -> 'd) -> ('e -> 'f) -> ('a, 'c, 'e) probability_constr -> ('b, 'd, 'f) probability_constr

val pp_probability_constr : 'iv 'tv 't. (Ppx_deriving_runtime.Format.formatter -> 'iv -> Ppx_deriving_runtime.unit) -> (Ppx_deriving_runtime.Format.formatter -> 'tv -> Ppx_deriving_runtime.unit) -> (Ppx_deriving_runtime.Format.formatter -> 't -> Ppx_deriving_runtime.unit) -> Ppx_deriving_runtime.Format.formatter -> ('iv, 'tv, 't) probability_constr -> Ppx_deriving_runtime.unit

val show_probability_constr : 'iv 'tv 't. (Ppx_deriving_runtime.Format.formatter -> 'iv -> Ppx_deriving_runtime.unit) -> (Ppx_deriving_runtime.Format.formatter -> 'tv -> Ppx_deriving_runtime.unit) -> (Ppx_deriving_runtime.Format.formatter -> 't -> Ppx_deriving_runtime.unit) -> ('iv, 'tv, 't) probability_constr -> Ppx_deriving_runtime.string

val fold_probability_constr : ('a -> 'b -> 'c) -> ('a -> 'd -> 'c) -> ('c -> 'e -> 'c) -> 'a -> ('b, 'd, 'e) probability_constr -> 'c

val name : ('a, 'b, 'c, 'd, 'e, 'f) clock_constr -> string

Clock constraint name.

val clocks : ('a, 'b, 'c, 'd, 'e, 'f) clock_constr -> 'a list

Lists all constraint clocks.

val map_time_const : ('a -> 'b) -> ('c -> 'd) -> ('e -> 'f) -> ('g, 'h, 'i, 'a, 'c, 'e) clock_constr -> ('g, 'h, 'i, 'b, 'd, 'f) clock_constr

Maps only the time constants in a logical constraint.