Module MicroStep.HardRelation

type 'id t = 'id * 'id

Type denoting relation x before y.

val t_of_sexp : 'id. (Sexplib0.Sexp.t -> 'id) -> Sexplib0.Sexp.t -> 'id t

val sexp_of_t : 'id. ('id -> Sexplib0.Sexp.t) -> 'id t -> Sexplib0.Sexp.t

val of_logical_spec : ('a, 'b, 'c, 'd, 'e, 'f) Mrtccsl__.Language.Cstr.clock_constr list -> 'a t list

val spec_of_sexp : (Sexplib0.Sexp.t -> 'a) -> Sexplib0__.Sexp.t -> 'a t list

val sexp_of_spec : ('a -> Sexplib0.Sexp.t) -> 'a t list -> Sexplib0__.Sexp.t

type 'a compiled = ('a, ('a, 'a) Either.t list) Prelude.Hashtbl.t

val compile : ('a * 'a) list -> 'a compiled

val arrange_randomly : 'a compiled -> 'a Prelude.Iter.t -> 'a Prelude.List.t

Orders present causaly-related clocks inside the label as indicated in restrictions, thus the information about causality should be supplied per step basis.