Module CausalEffect.EventEqTransition

Parameters

module T : sig ... end

module Time : Time

Signature

include sig ... end

module Time : sig ... end

module Place : sig ... end

Place IDs.

module Transition : sig ... end

Transition IDs.

module Color : sig ... end

Color IDs.

module Event : sig ... end

Event IDs.

module Probe : sig ... end

Probe IDs.

module Places : sig ... end

module Transitions : sig ... end

module Events : sig ... end

module Probes : sig ... end

module Coloring : sig ... end

module Instant : sig ... end

module Annot : sig ... end

module InstantSet : sig ... end

module Token : sig ... end

type place = Network(T)(Time).place =

| Variable of Token.t * bool

| Queue of Token.t list

Variable type.

val read_from_place : place -> Token.t option * place

val write_to_place : Token.t -> place -> place

type transition = Network(T)(Time).transition = {

input_arcs : Place.t list;

output_arcs : Place.t list;

coloring : Coloring.t;

event : Event.t;

}

type t = Network(T)(Time).t = {

places : place Places.t;

transitions : transition Transitions.t;

probes : (Coloring.guard * Transition.t) Probes.t;

}

Network type.

val empty : t

exception IncompatiblePlaceDefinition of Place.t

Raised when code declares that a place is both variable and queue at the same time.

exception MissingTransition of Transition.t

Raised when transition is expected, but not found.

exception IncompatibleTransitionDefinition of Transition.t

Raised when transition is declared twice with different events.

exception IncompatibleProbeDefinition of Probe.t

Raised when probe is declared twice with different matching colors or at different transition.

val add_variable : Places.key -> t -> t

val add_queue : Places.key -> t -> t

val add_transition : Transitions.key -> Event.t -> t -> t

val modify_transition : Transitions.key -> (transition -> transition) -> t -> t

val add_source : Transitions.key -> Place.t -> t -> t

val add_target : Transitions.key -> Place.t -> t -> t

val inject_color : Transitions.key -> Coloring.elt -> t -> t

val add_probe : Probes.key -> Coloring.guard -> Transition.t -> t -> t

module Compiled : sig ... end

Module for networks ready to travers traces efficiently. The representation is mutable.

val declare_data_connection : (Place.t -> 'a -> t) -> Place.t -> Transitions.key list -> Transitions.key list -> 'a -> t

Adds a place under a variable and arcs in transitions reading and writing to the place.

val sampling_variable : Place.t -> Transitions.key list -> Transitions.key list -> t -> t

Established an order-dependent one-to-many relation between write and read events.

val queueing : Place.t -> Transitions.key list -> Transitions.key list -> t -> t

Establishes a one-to-one relationship between the write and read events.

module Declaration : sig ... end

type end_semantics =

| Early
(*
Early denotes early cause or effect among equivalent.
*)

| Late
(*
Late denotes late cause or effect among equivalent.
*)

Semantics of ends of causal-effect chains.

val select_relevant_tokens : cause:end_semantics -> conseq:end_semantics -> Coloring.guard -> Token.mark Token.cause Prelude.Seq.t -> Token.mark Token.cause Prelude.Seq.t

Filters and maps tokens by selecting chain between the pair of cause and effect (reaction) in the equivalence class and the token.

exception IntervalMultipleCauses

When computing intervals between chains, each chain should have selected exactly one cause.

val full_interval : (unit -> 'a Token.cause Prelude.Seq.node) -> ('a * 'a Token.cause) Prelude.Seq.t

Adds duration of the interval between (succesuful) causal-effect chains (tokens).

raises IntervalMultipleCauses
when tokens were not pre-filtered to contain singular cause.