Module EventEqTransition.Token

type !'mark cause = 'mark Network(T)(Time).Token.cause =

| Internal
(*
Dependency on variable that was initialized to some value at the start of the system.
*)

| External of {

dependencies : 'mark cause list;
(*
Dependency cannot be empty.
*)

mark : 'mark;
(*
Event that binds the dependencies into a new relation. Cannot be empty.
*)

}
(*
Value depends on external activity denoted by the 'mark.
*)

Definition of influence relation for marks of type 'mark.

val compare_cause : ('mark -> 'mark -> int) -> 'mark cause -> 'mark cause -> int

val cause_of_sexp : (Sexplib0.Sexp.t -> 'mark) -> Sexplib0.Sexp.t -> 'mark cause

val sexp_of_cause : ('mark -> Sexplib0.Sexp.t) -> 'mark cause -> Sexplib0.Sexp.t

val fold_cause : ('a -> 'b -> 'a) -> 'a -> 'b cause -> 'a

type mark = Network(T)(Time).Token.mark = {

instant : Instant.t;

annotation : Annot.t;

}

val mark_of_sexp : Sexplib0.Sexp.t -> mark

val sexp_of_mark : mark -> Sexplib0.Sexp.t

val compare_mark : mark -> mark -> int

type t = mark cause

Token type. Represents data passing mark a system. Witnesses the relationships between events from the data interactions. Has an arbitrary combined categorization (color).

val t_of_sexp : Sexplib0.Sexp.t -> t

val sexp_of_t : t -> Sexplib0.Sexp.t

val fold : 'a -> 'b -> 'a

val compare : t -> t -> int

val root_mark_opt : 'a cause -> 'a option

exception NoRootMark

val root_mark : 'a cause -> 'a

val non_transitive_causes : 'a cause -> 'a list

val colors : mark cause -> Coloring.t

val internal : 'a cause

val dependencies_annotation : mark cause list -> Annot.t

val build_external : Instant.t -> Coloring.t -> mark cause Prelude.List.t -> mark cause

val contains_internal : 'a cause -> bool

val is_internal_free : 'a cause -> bool

val chromatic_filter : Coloring.guard -> mark cause -> mark cause option

val unique_instants : mark cause -> InstantSet.t

val has_common_instants : mark cause -> mark cause -> bool

val earliest_latest_conseq : minmax:((mark cause -> mark cause -> int) -> 'a -> 'b) -> 'a -> 'b

val subcause : dep_select: (('a option * 'b cause list) -> 'b cause -> 'a option * 'b cause list) -> 'b cause -> 'b cause

val early_cause_select : (Time.t option * mark cause list) -> mark cause -> Time.t option * mark cause list

val late_cause_select : (Time.t option * mark cause list) -> mark cause -> Time.t option * mark cause list