Module CausalEffect.Network

Functor that provides construction and traversal of causal (through data sharing) networks.

Parameters

module IDs : IDs

module Time : Time

Signature

include module type of struct include IDs end

module Place = IDs.Place

Place IDs.

module Transition = IDs.Transition

Transition IDs.

module Color = IDs.Color

Color IDs.

module Event = IDs.Event

Event IDs.

module Probe = IDs.Probe

Probe IDs.

module Places : sig ... end

module Transitions : sig ... end

module Events : sig ... end

module Probes : sig ... end

type place_type =

| Variable

| Queue

module Coloring : sig ... end

type token = {

visits : Time.t Prelude.Dynarray.t Events.t;

coloring : Coloring.t;

}

Token type. Represents data passing through a system. Witnesses the relationships between events from the data interactions.

val compare_token : token -> token -> int

val token_of_sexp : Sexplib0.Sexp.t -> token

val sexp_of_token : token -> Sexplib0.Sexp.t

val empty_token : token

An empty token. Should be interpretted as incorrect data because is unrelated to any external event.

val merge_tokens : token -> token -> token

val add_visit : token -> Events.key -> Time.t -> Coloring.t -> token

type inner_place =

| Variable of token * bool

| Queue of token list

Variable type.

val read_from_place : inner_place -> token option * inner_place

val write_to_place : token -> inner_place -> inner_place

type inner_transition = {

input_arcs : Place.t list;

output_arcs : Place.t list;

coloring : Coloring.t;

event : Event.t;

}

type color_match =

| Any

| Single of Color.t

| And of color_match * color_match

| Or of color_match * color_match

Type of color matches.

val compare_color_match : color_match -> color_match -> int

val eval_color_match : Stdlib__Set.Make(Color).t -> color_match -> bool

type t = {

places : inner_place Places.t;

transitions : inner_transition Transitions.t;

probes : (color_match * Transition.t) Probes.t;

}

Network type.

val empty : t

Empty network.

exception IncompatiblePlaceDefinition of Place.t

Raised when code declares that a place is both variable and queue at the same time.

exception MissingTransition of Transition.t

Raised when transition is expected, but not found.

exception IncompatibleTransitionDefinition of Transition.t

Raised when transition is declared twice with different events.

exception IncompatibleProbeDefinition of Probe.t

Raised when probe is declared twice with different matching colors or at different transition.

val add_variable : Places.key -> t -> t

Adds variable place to the network.

val add_queue : Places.key -> t -> t

Adds queue variable to the network.

val add_transition : Transitions.key -> Event.t -> t -> t

Adds an empty transition with an event.

val modify_transition : Transitions.key -> (inner_transition -> inner_transition) -> t -> t

val add_source : Transitions.key -> Place.t -> t -> t

Adds input arc (dependency) to a transition. All inputs need to have data in order to activate the transition.

val add_target : Transitions.key -> Place.t -> t -> t

Adds output arc (cause) to a transition.

val inject_color : Transitions.key -> Coloring.elt -> t -> t

Adds coloring to output tokens of a transition.

val add_probe : Probes.key -> color_match -> Transition.t -> t -> t

Adds a probe on output tokens of the transition.

module Compiled : sig ... end

Module for networks ready to travers traces efficiently. The representation is mutable.