Module Functional_chain.Make

Parameters

module C : Mrtccsl.Backend.Naive.Hashed.ID

module N : sig ... end

Signature

module A : sig ... end

module ST : sig ... end

module CMap : sig ... end

module Trace : sig ... end

type chain_instance = {

trace : N.t CMap.t;

misses : int CMap.t;

since : N.t option;

}

val map_chain_instance : chain_instance -> chain_instance

type partial_chain = {

trace : N.t CMap.t;

targets : int CMap.t;

mutable misses : int CMap.t;

since : N.t option;

}

val partial_chain_to_string : partial_chain -> string

type semantics =

| All

| Earliest

| Latest

| Randomized

val consume_label : ?sem:semantics -> update_tick:(BitvectorSet.BijectionToInt.Hashed(C).elt -> N.t -> N.t) -> (BitvectorSet.BijectionToInt.Hashed(C).elt * [< `Causality | `New | `NewPeriodic | `Sampling Causality Sampling ]) array -> (chain_instance Prelude.Dynarray.t * partial_chain Prelude.Queue.t array * int array) -> (Fast_bitvector.t, N.t) Trace.step -> chain_instance Prelude.Dynarray.t * partial_chain Prelude.Queue.t array * int array

val extract_chains : semantics -> BitvectorSet.BijectionToInt.Hashed(C).elt Chain.t -> (Fast_bitvector.t, N.t) Trace.step Prelude.Seq.t -> chain_instance Prelude.Dynarray.t * partial_chain Prelude.Queue.t array

val chains_of_spec : ?debug:bool -> ?sem:semantics -> (A.sol_strategy * int * N.t) -> (BitvectorSet.BijectionToInt.Hashed(C).elt, A.CMap.key, A.C.t, A.CMap.key, A.CMap.key, A.NI.num) Language.Specification.t -> C.t Chain.t -> (A.L.t, A.N.t) Trace.step Prelude.Seq.t * bool * chain_instance Prelude.Dynarray.t * partial_chain Prelude.Queue.t array

val reaction_time_of_span : chain_instance -> (CMap.key * CMap.key) -> N.t

val statistics : CMap.key -> chain_instance Prelude.Iter.t -> (Prelude.Int.t * int) Prelude.Iter.t

val print_statistics : CMap.key -> Stdlib.Format.formatter -> chain_instance Prelude.Iter.t -> unit

val categorized_reaction_times : CMap.key list -> (CMap.key * CMap.key) -> chain_instance Prelude.Seq.t -> (string * N.t) Prelude.Seq.t

val weighted_full_reaction_times : ?discretize_range:(N.t -> N.t Prelude.Seq.t) -> CMap.key Chain.t -> chain_instance Prelude.Seq.t -> ((string * N.t) list * N.t) Prelude.Seq.t

weighted_full_reaction_times ~discretize_range chain chain_instances returns end-to-end reaction time of the functional chain using ~discretize_range to randomly offset the reaction times, inside (0, start[i+1]-start[i]) range.

val reaction_times_to_string : sep:string -> ('a * (C.t * C.t, N.t) Prelude.Hashtbl.t) Stdlib.Seq.t -> string

val reaction_times_to_csv : CMap.key list -> (CMap.key * CMap.key) list -> chain_instance Prelude.Seq.t -> Stdlib.out_channel -> unit