Module Specification.Builder

Helper module for constructing specifications in imperative style.

type ('c, 'tp, 'ip, 'tv, 'iv, 't) spec = ('c, 'tp, 'ip, 'tv, 'iv, 't) t

Reexport of the specification type.

type nonrec ('c, 'tp, 'ip, 'tv, 'iv, 't) t = {

mutable spec : ('c, 'tp, 'ip, 'tv, 'iv, 't) t;

}

Builder type.

val make : unit -> ('a, 'b, 'c, 'd, 'e, 'f) t

Creates a builder with empty specification.

val logical : ('a, 'b, 'c, 'd, 'e, 'f) t -> ('a, 'b, 'c, 'd, 'e, 'f) Cstr.clock_constr -> unit

Adds logical constraint to the specification.

val duration : ('a, 'b, 'c, 'd, 'e, 'f) t -> ('d, 'f) Cstr.numeric_rel_constr -> unit

Adds time duration relation constraint to the specification.

val probab : ('a, 'b, 'c, 'd, 'e, 'f) t -> ('e, 'd, 'f) Cstr.probability_constr -> unit

Adds probabilistic constraint to the specification.

val integer : ('a, 'b, 'c, 'd, 'e, 'f) t -> ('e, int) Cstr.numeric_rel_constr -> unit

Adds integer relation constraint to the specification.

val get : ('a, 'b, 'c, 'd, 'e, 'f) t -> ('a, 'b, 'c, 'd, 'e, 'f) t

Returns specification with the added constraints in the push order.

type ('c, 'tp, 'ip, 'tv, 'iv, 't) decl = ('c, 'tp, 'ip, 'tv, 'iv, 't) t -> unit

Declaration is a function that adds the builder with constraints.

val empty_decl : 'a -> unit

val of_decl : ('a, 'b, 'c, 'd, 'e, 'f) decl -> ('a, 'b, 'c, 'd, 'e, 'f) spec

Returns the specification defined using a declaration function.

type ('s, 'c, 'tp, 'ip, 'tv, 'iv, 't) stateful_decl = 's -> ('c, 'tp, 'ip, 'tv, 'iv, 't) t -> 's

Stateful declaration is a function that adds the builder with constraints while capturing and returning some state.

val of_stateful_decl : 'a -> ('a, 'b, 'c, 'd, 'e, 'f, 'g) stateful_decl -> 'a * ('b, 'c, 'd, 'e, 'f, 'g) spec