Module MakeIO.Serialize

val undefined : L.t -> (L.E.t -> unit) -> unit

undefined serializes the step in order defined by the label implementation.

val random : L.t -> L.E.t Prelude.Iter.t

random serializes the step in randomly.

val respect_microstep : (L.E.t, (L.E.t, L.E.t) Either.t list) Prelude.Hashtbl.t -> L.t -> L.E.t Prelude.Iter.t

respect_microstep serializes the step randomly while respecting partial ordering.