Module Prelude

val (<<) : ('a -> 'b) -> ('c -> 'a) -> 'c -> 'b

Right-to-left function composition, f << g if the same as f(g(x)).

val (>>) : ('a -> 'b) -> ('b -> 'c) -> 'a -> 'c

Left-to-right function composition, g >> f if the same as f(g(x)).

val (>==) : 'a option -> ('a -> 'b option) -> 'b option

val (let*) : 'a option -> ('a -> 'b option) -> 'b option

Option let-bind syntax.

val (and*) : 'a option -> 'b option -> ('a * 'b) option

Option and-bind syntax.

val failwithf : ('a, unit, string, 'b) Stdlib.format4 -> 'a

Raise exception with the given format string and parameters.

module Fun : sig ... end

module Int : sig ... end

module Option : sig ... end

module Result : sig ... end

module Seq : sig ... end

module List : sig ... end

module String : sig ... end

module Buffer : sig ... end

module Tuple : sig ... end

module Hashtbl : sig ... end

module Map : sig ... end

module Set : sig ... end

val fixpoint : ('a -> 'a -> bool) -> ('a -> 'a) -> 'a -> 'a

fixpoint init eq f start returns a value v such that f(v) = v by starting from v.

module CircularList : sig ... end

module Array : sig ... end

module Dynarray : sig ... end

module Iter : sig ... end

module Queue : sig ... end

module Heap (Elem : Map.OrderedType) : sig ... end

module In_channel : sig ... end

module Sys : sig ... end