Module Prelude.Seq

include module type of struct include Stdlib.Seq end

type !'a t = unit -> 'a node

and !'a node = 'a Stdlib__Seq.node =

| Nil

| Cons of 'a * 'a t

val is_empty : 'a t -> bool

val uncons : 'a t -> ('a * 'a t) option

val length : 'a t -> int

val iter : ('a -> unit) -> 'a t -> unit

val fold_left : ('acc -> 'a -> 'acc) -> 'acc -> 'a t -> 'acc

val iteri : (int -> 'a -> unit) -> 'a t -> unit

val fold_lefti : ('acc -> int -> 'a -> 'acc) -> 'acc -> 'a t -> 'acc

val for_all : ('a -> bool) -> 'a t -> bool

val exists : ('a -> bool) -> 'a t -> bool

val find : ('a -> bool) -> 'a t -> 'a option

val find_index : ('a -> bool) -> 'a t -> int option

val find_map : ('a -> 'b option) -> 'a t -> 'b option

val find_mapi : (int -> 'a -> 'b option) -> 'a t -> 'b option

val iter2 : ('a -> 'b -> unit) -> 'a t -> 'b t -> unit

val fold_left2 : ('acc -> 'a -> 'b -> 'acc) -> 'acc -> 'a t -> 'b t -> 'acc

val for_all2 : ('a -> 'b -> bool) -> 'a t -> 'b t -> bool

val exists2 : ('a -> 'b -> bool) -> 'a t -> 'b t -> bool

val equal : ('a -> 'b -> bool) -> 'a t -> 'b t -> bool

val compare : ('a -> 'b -> int) -> 'a t -> 'b t -> int

val empty : 'a t

val return : 'a -> 'a t

val cons : 'a -> 'a t -> 'a t

val singleton : 'a -> 'a t

val init : int -> (int -> 'a) -> 'a t

val unfold : ('b -> ('a * 'b) option) -> 'b -> 'a t

val repeat : 'a -> 'a t

val forever : (unit -> 'a) -> 'a t

val cycle : 'a t -> 'a t

val iterate : ('a -> 'a) -> 'a -> 'a t

val map : ('a -> 'b) -> 'a t -> 'b t

val mapi : (int -> 'a -> 'b) -> 'a t -> 'b t

val filter : ('a -> bool) -> 'a t -> 'a t

val filteri : (int -> 'a -> bool) -> 'a t -> 'a t

val filter_map : ('a -> 'b option) -> 'a t -> 'b t

val scan : ('b -> 'a -> 'b) -> 'b -> 'a t -> 'b t

val take : int -> 'a t -> 'a t

val drop : int -> 'a t -> 'a t

val take_while : ('a -> bool) -> 'a t -> 'a t

val drop_while : ('a -> bool) -> 'a t -> 'a t

val group : ('a -> 'a -> bool) -> 'a t -> 'a t t

val memoize : 'a t -> 'a t

exception Forced_twice

val once : 'a t -> 'a t

val transpose : 'a t t -> 'a t t

val append : 'a t -> 'a t -> 'a t

val concat : 'a t t -> 'a t

val flat_map : ('a -> 'b t) -> 'a t -> 'b t

val concat_map : ('a -> 'b t) -> 'a t -> 'b t

val zip : 'a t -> 'b t -> ('a * 'b) t

val map2 : ('a -> 'b -> 'c) -> 'a t -> 'b t -> 'c t

val interleave : 'a t -> 'a t -> 'a t

val sorted_merge : ('a -> 'a -> int) -> 'a t -> 'a t -> 'a t

val product : 'a t -> 'b t -> ('a * 'b) t

val map_product : ('a -> 'b -> 'c) -> 'a t -> 'b t -> 'c t

val unzip : ('a * 'b) t -> 'a t * 'b t

val split : ('a * 'b) t -> 'a t * 'b t

val partition_map : ('a -> ('b, 'c) Stdlib.Either.t) -> 'a t -> 'b t * 'c t

val partition : ('a -> bool) -> 'a t -> 'a t * 'a t

val of_dispenser : (unit -> 'a option) -> 'a t

val to_dispenser : 'a t -> unit -> 'a option

val ints : int -> int t

val int_seq : ?start:int -> int -> int t

int_seq ~start n returns sequence start...start+n (not included).

parameter start
defaults to 0

val int_seq_inclusive : (int * int) -> int t

int_seq_inclusive (x,y) returns sequence x..=y.

val zip_list : 'a Stdlib.Seq.t list -> 'a list t

Returns sequence of lists from list of sequences. Ends when any sequence ends.

val fold_left_until : ('a -> bool) -> ('b -> 'a -> 'b) -> 'b -> 'a t -> 'b * 'a option * 'a t

Folds the sequence until predicate p is not satisfied.

returns
accumulator, first non-satisfied element and the rest of the sequence.

val return2 : 'a -> 'a -> 'a t

Returns sequence of x followed by y.

val to_string : ?sep:string -> ('a -> string) -> 'a Stdlib.Seq.t -> string

Returns string representation of the sequence given to_string of the elements.

val unfold_for_while : ('a -> ('b * 'a) option) -> 'a -> int -> ('b -> bool) -> 'b Stdlib.Seq.t * bool Stdlib.ref

Returns an unfolding sequences seeded with init state, until predicate p is satisfied on that sequence or until the sequence length is less than n.

returns
the sequence and bool reference that indicates that the sequences was cut according to the conditions.

val append_list : 'a t list -> 'a t

Appends all sequences in the list.

val reduce_left : ('acc -> 'a -> 'acc) -> ('a -> 'acc) -> 'a t -> 'acc

reduce_left folds the sequence just as fold_left, except it uses the first element as as its accumulator.

val of_pair : ('a * 'a) -> 'a t

Converts pait into a sequence, left element is first.

val product_seq : 'a t t -> 'a t t

Returns a sequence of cartesian products between the input sequences.

val uncons_exn : 'a t -> 'a * 'a t

Returns the first element and rest of the sequence. Raises exception when empty.

val uncons_opt : (unit -> 'a node) -> 'a option * 'a t

val to_tuple2 : 'a t -> 'a * 'a

Unpacks sequence into a pair.

val to_tuple3 : 'a t -> 'a * 'a * 'a

Unpacks sequence into a triple.

val to_tuple4 : 'a t -> 'a * 'a * 'a * 'a

Unpacks sequence into a 4-element tuple.

val to_tuple5 : 'a t -> 'a * 'a * 'a * 'a * 'a

Unpacks sequence into a 5-element tuple.

val pp : ?sep:string -> (Stdlib.Format.formatter -> 'a -> unit) -> Stdlib.Format.formatter -> 'a Stdlib.Seq.t -> unit

Pretty prints a sequence into a formatter.

val fold_leftir_aux : ('acc -> int -> 'a -> ('acc, 'b) Stdlib.result) -> ('acc, 'b) Stdlib.result -> int -> 'a t -> ('acc, 'b) Stdlib.result

val fold_leftir : ('a -> int -> 'b -> ('a, 'c) Stdlib.result) -> ('a, 'c) Stdlib.result -> 'b t -> ('a, 'c) Stdlib.result

Folds an indexed sequence with a Result accumulator.

val fold_leftr : ('a -> 'b -> ('a, 'c) Stdlib.result) -> ('a, 'c) Stdlib.result -> 'b t -> ('a, 'c) Stdlib.result

Folds a sequence with a Result accumulator.

val minmax : ('a -> 'a -> int) -> 'a t -> ('a * 'a) option