Module STS.Interpretation

Module of value based interpretation of the abstract machine.

type var = string

type state = {

integers : int VarMap.t;

rationals : Common.Number.Rational.t VarMap.t;

bools : bool VarMap.t;

int_queues : int Queue.t VarMap.t;

rat_queues : Common.Number.Rational.t Queue.t VarMap.t;

}

Type of state values.

val state_of_sexp : Sexplib0.Sexp.t -> state

val sexp_of_state : state -> Sexplib0.Sexp.t

val default_state : state

Default (unassigned/empty) state.

type ('v, 'a) get_value = 'v -> 'a

type 'v state_interface = {

integer : ('v, int) get_value;

rational : ('v, Common.Number.Rational.t) get_value;

bool : ('v, bool) get_value;

int_queue : ('v, int Queue.t) get_value;

rat_queue : ('v, Common.Number.Rational.t Queue.t) get_value;

}

val state_to_interface : state -> VarMap.key state_interface

type inputs = {

rationals : Common.Number.Rational.t VarMap.t;

integers : int VarMap.t;

bools : bool VarMap.t;

}

Type of input values.

val default_inputs : inputs

type ('v, 'i, 'r) input_interface = {

integer : ('v, 'i) get_value;

rational : ('v, 'r) get_value;

bool : ('v, bool) get_value;

}

val empty_input_interface : (empty, 'a, 'b) input_interface

module Full : sig ... end

val apply_assignment : var state_interface -> (var, int, Common.Number.Rational.t) input_interface -> state -> (VarMap.key * (var, var) expr) -> state

Applies the assignment, evaluating on the given state and input values, and saves the value in new state.

val apply_assignments : var state_interface -> (var, int, Common.Number.Rational.t) input_interface -> state -> (VarMap.key * (var, var) expr) list -> state

Applies all assignments, evaluating on the given state and input values, and saves values in new state.

module Partial : sig ... end

type transition_error =

| NoValidTransition

| FailedGuard

| FailedInvariant

val transition_error_to_string : transition_error -> string

val do_transition : input_to_interface:(inputs -> 'a) -> eval_guard: (VarMap.key state_interface -> 'a -> (VarMap.key, var) bool_expr -> bool) -> (VarMap.key, var) t -> state -> inputs -> (state, transition_error) Stdlib.result

val apply_transition : (VarMap.key, var) t -> state -> inputs -> (state, transition_error) Stdlib.result

Full evaluate the step in a transition, expects all variables to have defined values.

val accept_transition : (VarMap.key, var) t -> state -> inputs -> (state, transition_error) Stdlib.result

Evaluate the step in a transition, while existentially quantifying the numerical variables. Allows to check specification satisfaction on traces with only timestampts and clocks.