Module Interpretation.Partial

module Ident : sig ... end

module TermQ : sig ... end

module TermZ : sig ... end

module DZ : sig ... end

module DQ : sig ... end

val input_to_interface : inputs -> (Ident.t, Vpl.Domains.UncertifiedZ.t, Vpl.Domains.UncertifiedQ.t) input_interface

module DQZ : sig ... end

module type Conj = sig ... end

module Disj (Conj : Conj) : sig ... end

module DQZF : sig ... end

Module of propositions on Q x Z.

module Dual : sig ... end

Module for partial Boolean function evaluation.

val get_iqueue : 'a state_interface -> 'a -> int Queue.t

val get_rqueue : 'a state_interface -> 'a -> Common.Number.Rational.t Queue.t

val eval_bool : 'sv state_interface -> (Ident.t, DZ.t, DQ.t) input_interface -> ('sv, Ident.t) bool_expr -> Dual.bool_result

Paritally evaluate Boolean formula given state and input values. Returns conditions of when the function returns true and false.

val eval_integer : 'sv state_interface -> (Ident.t, DZ.t, DQ.t) input_interface -> ('sv, Ident.t) int_expr -> Dual.int_result

Evaluate integer expression given state and input values.

val eval_rational : 'sv state_interface -> (Ident.t, DZ.t, DQ.t) input_interface -> ('sv, Ident.t) rat_expr -> Dual.rat_result

Evaluate rational expression given state and input values.