Module Partial.Dual

Module for partial Boolean function evaluation.

type bool_result = DQZF.t * DQZF.t

Type representing conditions on numerical variables of when Boolean function returns true and false.

val wrap_const_bool : bool -> DQZF.t * DQZF.t

val not : ('a * 'b) -> 'b * 'a

val equal : (DQZ.t list * DQZ.t list) -> (DQZ.t list * DQZ.t list) -> DQZF.t * DQZF.t

val (&&) : (DQZ.t list * DQZ.t list) -> (DQZ.t list * DQZ.t list) -> bool_result

val (||) : (DQZ.t list * DQZ.t list) -> (DQZ.t list * DQZ.t list) -> bool_result

val all : bool_result Common.Prelude.List.t -> bool_result

val any : bool_result Common.Prelude.List.t -> bool_result

type 'e result = (DQZF.t * 'e) list

Type of numerical function returns, conditioned with a proposition on variables.

type int_result = DZ.a_expr result

type rat_result = DQ.a_expr result

val do_rel : of_relation:(('a * Vpl.Cstr_type.cmpT_extended * 'a) -> 'b) -> apply:('b -> DQZ.t -> DQZ.t) -> 'a result -> 'a result -> [< `Eq | `Less | `LessEq | `More | `MoreEq | `Neq ] -> bool_result

Constructs pair of conditions for when the relation is satisfied and not.

val int_do_rel : int_result -> int_result -> [< `Eq | `Less | `LessEq | `More | `MoreEq | `Neq ] -> bool_result

val rat_do_rel : rat_result -> rat_result -> [< `Eq | `Less | `LessEq | `More | `MoreEq | `Neq ] -> bool_result

val wrap_const_result : 'a -> (DQZF.t * 'a) list

When expression is not conditioned with a predicate, such as constant.