Module Symbolic.Machine

module VarMap : sig ... end

Map from variables to (typically) their values.

module Queue : sig ... end

type ('bool, 'expr) ite = {

cond : 'bool;
(*
If guard.
*)

if_true : 'expr;
(*
To be evaluated when the condition is true.
*)

if_false : 'expr;
(*
To be evaluated when the condition is false.
*)

}

Generic if-then-else type.

val compare_ite : 'bool 'expr. ('bool -> 'bool -> int) -> ('expr -> 'expr -> int) -> ('bool, 'expr) ite -> ('bool, 'expr) ite -> int

val ite_of_sexp : 'bool 'expr. (Sexplib0.Sexp.t -> 'bool) -> (Sexplib0.Sexp.t -> 'expr) -> Sexplib0.Sexp.t -> ('bool, 'expr) ite

val sexp_of_ite : 'bool 'expr. ('bool -> Sexplib0.Sexp.t) -> ('expr -> Sexplib0.Sexp.t) -> ('bool, 'expr) ite -> Sexplib0.Sexp.t

type empty = |

Empty type. It has no constructors and so values cannot exist. Used to prune possibilities in polymorphic types.

type ('sv, 'iv) bool_expr =

| BConst of bool

| BStateVar of 'sv

| BInputVar of 'iv

| BNot of ('sv, 'iv) bool_expr

| BAnd of ('sv, 'iv) bool_expr list

| BOr of ('sv, 'iv) bool_expr list

| BEq of ('sv, 'iv) bool_expr * ('sv, 'iv) bool_expr

| BNeq of ('sv, 'iv) bool_expr * ('sv, 'iv) bool_expr

| BImply of ('sv, 'iv) bool_expr * ('sv, 'iv) bool_expr

| IntComp of ('sv, 'iv) int_expr * Expr.num_rel * ('sv, 'iv) int_expr

| RatComp of ('sv, 'iv) rat_expr * Expr.num_rel * ('sv, 'iv) rat_expr

| BITE of (('sv, 'iv) bool_expr, ('sv, 'iv) bool_expr) ite

| IntQueuePositive of 'sv

Type of Boolean expressions.

and ('sv, 'iv) int_expr =

| IConst of int

| IStateVar of 'sv

| IInputVar of 'iv

| IBinOp of ('sv, 'iv) int_expr * Expr.num_op * ('sv, 'iv) int_expr

| IITE of (('sv, 'iv) bool_expr, ('sv, 'iv) int_expr) ite

| IPeekFirstQueue of 'sv

| IPeekLastQueue of 'sv

| IntQueueLength of 'sv

| RatQueueLength of 'sv

Type of integer expressions.

and ('sv, 'iv) rat_expr =

| RConst of Number.Rational.t

| RStateVar of 'sv

| RInputVar of 'iv

| RITE of (('sv, 'iv) bool_expr, ('sv, 'iv) rat_expr) ite

| RBinOp of ('sv, 'iv) rat_expr * Expr.num_op * ('sv, 'iv) rat_expr

| RPeekFirstQueue of 'sv

| RPeekLastQueue of 'sv

Type of rational expressions.

val compare_bool_expr : 'sv 'iv. ('sv -> 'sv -> int) -> ('iv -> 'iv -> int) -> ('sv, 'iv) bool_expr -> ('sv, 'iv) bool_expr -> int

val compare_int_expr : 'sv 'iv. ('sv -> 'sv -> int) -> ('iv -> 'iv -> int) -> ('sv, 'iv) int_expr -> ('sv, 'iv) int_expr -> int

val compare_rat_expr : 'sv 'iv. ('sv -> 'sv -> int) -> ('iv -> 'iv -> int) -> ('sv, 'iv) rat_expr -> ('sv, 'iv) rat_expr -> int

val bool_expr_of_sexp : 'sv 'iv. (Sexplib0.Sexp.t -> 'sv) -> (Sexplib0.Sexp.t -> 'iv) -> Sexplib0.Sexp.t -> ('sv, 'iv) bool_expr

val int_expr_of_sexp : 'sv 'iv. (Sexplib0.Sexp.t -> 'sv) -> (Sexplib0.Sexp.t -> 'iv) -> Sexplib0.Sexp.t -> ('sv, 'iv) int_expr

val rat_expr_of_sexp : 'sv 'iv. (Sexplib0.Sexp.t -> 'sv) -> (Sexplib0.Sexp.t -> 'iv) -> Sexplib0.Sexp.t -> ('sv, 'iv) rat_expr

val sexp_of_bool_expr : 'sv 'iv. ('sv -> Sexplib0.Sexp.t) -> ('iv -> Sexplib0.Sexp.t) -> ('sv, 'iv) bool_expr -> Sexplib0.Sexp.t

val sexp_of_int_expr : 'sv 'iv. ('sv -> Sexplib0.Sexp.t) -> ('iv -> Sexplib0.Sexp.t) -> ('sv, 'iv) int_expr -> Sexplib0.Sexp.t

val sexp_of_rat_expr : 'sv 'iv. ('sv -> Sexplib0.Sexp.t) -> ('iv -> Sexplib0.Sexp.t) -> ('sv, 'iv) rat_expr -> Sexplib0.Sexp.t

type ('sv, 'iv) int_queue_expr =

| IQVar of 'sv

| IPushQueue of ('sv, 'iv) int_queue_expr * ('sv, 'iv) int_expr

| IPopQueue of ('sv, 'iv) int_queue_expr

| IDecreaseAllQueue of ('sv, 'iv) int_queue_expr

| IIncreaseAllQueue of ('sv, 'iv) int_queue_expr

| IQITE of (('sv, 'iv) bool_expr, ('sv, 'iv) int_queue_expr) ite

Type of integer queue expressions.

val compare_int_queue_expr : 'sv 'iv. ('sv -> 'sv -> int) -> ('iv -> 'iv -> int) -> ('sv, 'iv) int_queue_expr -> ('sv, 'iv) int_queue_expr -> int

type ('sv, 'iv) rat_queue_expr =

| RQVar of 'sv

| RPushQueue of ('sv, 'iv) rat_queue_expr * ('sv, 'iv) rat_expr

| RPopQueue of ('sv, 'iv) rat_queue_expr

| RQITE of (('sv, 'iv) bool_expr, ('sv, 'iv) rat_queue_expr) ite

Type of rational queue expressions.

val compare_rat_queue_expr : 'sv 'iv. ('sv -> 'sv -> int) -> ('iv -> 'iv -> int) -> ('sv, 'iv) rat_queue_expr -> ('sv, 'iv) rat_queue_expr -> int

type ('sv, 'iv) expr =

| BoolExpr of ('sv, 'iv) bool_expr

| IntExpr of ('sv, 'iv) int_expr

| RatExpr of ('sv, 'iv) rat_expr

| IntQueueExpr of ('sv, 'iv) int_queue_expr

| RatQueueExpr of ('sv, 'iv) rat_queue_expr

Type of expressions.

val compare_expr : 'sv 'iv. ('sv -> 'sv -> int) -> ('iv -> 'iv -> int) -> ('sv, 'iv) expr -> ('sv, 'iv) expr -> int

type ('sv, 'iv) guard = ('sv, 'iv) bool_expr

Type of transition guards.

val compare_guard : 'sv 'iv. ('sv -> 'sv -> int) -> ('iv -> 'iv -> int) -> ('sv, 'iv) guard -> ('sv, 'iv) guard -> int

type ('sv, 'iv) assignment = 'sv * ('sv, 'iv) expr

Type of transition assignments.

val compare_assignment : 'sv 'iv. ('sv -> 'sv -> int) -> ('iv -> 'iv -> int) -> ('sv, 'iv) assignment -> ('sv, 'iv) assignment -> int

type ('sv, 'iv) transition = {

state_cond : ('sv, empty) bool_expr;
(*
Condition on state variables that identifies the partition.
*)

input_cond : ('sv, 'iv) bool_expr;
(*
Input condition of the form S -> 2^(B^n * Q), where B^n encodes the clock ticks and Q encodes next possible time.
*)

assignments : ('sv, 'iv) assignment list;
(*
List of actions performed on the state variables using previous state and inputs S * B^n * Q -> S
*)

}

Transition type. Consists of conditions on state (preselect), inputs and state, and state update.

val sync_transitions : ('a, 'b) transition -> ('a, 'b) transition -> ('a, 'b) transition

Synchronizes two transitions into a new transition. A conjunction of conditions, union of assignments.

type ('sv, 'iv) t = {

transitions : ('sv, 'iv) transition list;

invariant : ('sv, empty) bool_expr;

}

Type of abstract machines.

val sync_machines : ('a, 'b) t -> ('a, 'b) t -> ('a, 'b) t

Synchronizes two machines by cartesian product of their transitions and conjunction of invariants.

module Interpretation : sig ... end

Module of value based interpretation of the abstract machine.

module Syntax : sig ... end

Syntax for the machine creation.