Module Interpretation.Full

val get_iqueue : 'a state_interface -> 'a -> int Queue.t

val get_rqueue : 'a state_interface -> 'a -> Number.Rational.t Queue.t

val input_to_interface : inputs -> (VarMap.key, int, Mpq.f Number.Rational.tt) input_interface

val eval_bool : 'sv state_interface -> ('iv, int, Number.Rational.t) input_interface -> ('sv, 'iv) bool_expr -> bool

Evaluate Boolean formula given state and input values.

val eval_integer : 'sv state_interface -> ('iv, int, Number.Rational.t) input_interface -> ('sv, 'iv) int_expr -> int

Evaluate integer expression given state and input values.

val eval_rational : 'sv state_interface -> ('iv, int, Number.Rational.t) input_interface -> ('sv, 'iv) rat_expr -> Number.Rational.t

Evaluate rational expression given state and input values.

val eval_int_queue : 'a state_interface -> ('b, int, Number.Rational.t) input_interface -> ('a, 'b) int_queue_expr -> int Queue.t

Evaluate integer queue expression given state and input values.

val eval_rat_queue : 'a state_interface -> ('b, int, Number.Rational.t) input_interface -> ('a, 'b) rat_queue_expr -> Number.Rational.t Queue.t

Evaluate rational queue expression given state and input values.