Module Machine.Syntax

Syntax for the machine creation.

Integers

val (>) : ('a, 'b) int_expr -> ('a, 'b) int_expr -> ('a, 'b) bool_expr

val (<) : ('a, 'b) int_expr -> ('a, 'b) int_expr -> ('a, 'b) bool_expr

val (>=) : ('a, 'b) int_expr -> ('a, 'b) int_expr -> ('a, 'b) bool_expr

val (<=) : ('a, 'b) int_expr -> ('a, 'b) int_expr -> ('a, 'b) bool_expr

val (==) : ('a, 'b) int_expr -> ('a, 'b) int_expr -> ('a, 'b) bool_expr

val (!=) : ('a, 'b) int_expr -> ('a, 'b) int_expr -> ('a, 'b) bool_expr

val iconst : int -> ('a, 'b) int_expr

val (+) : ('a, 'b) int_expr -> ('a, 'b) int_expr -> ('a, 'b) int_expr

val (-) : ('a, 'b) int_expr -> ('a, 'b) int_expr -> ('a, 'b) int_expr

val (*) : ('a, 'b) int_expr -> ('a, 'b) int_expr -> ('a, 'b) int_expr

val (/) : ('a, 'b) int_expr -> ('a, 'b) int_expr -> ('a, 'b) int_expr

val iite : ('a, 'b) bool_expr -> ('a, 'b) int_expr -> ('a, 'b) int_expr -> ('a, 'b) int_expr

val iinvar : 'a -> ('b, 'a) int_expr

val iitec : ('a, 'b) bool_expr -> int -> int -> ('a, 'b) int_expr

val ilength : 'a -> ('a, 'b) int_expr

val rlength : 'a -> ('a, 'b) int_expr

val i0 : ('a, 'b) int_expr

val i1 : ('a, 'b) int_expr

Rationals

val rinvar : 'a -> ('b, 'a) rat_expr

val rsvar : 'a -> ('a, 'b) rat_expr

val (>.) : ('a, 'b) rat_expr -> ('a, 'b) rat_expr -> ('a, 'b) bool_expr

val (<.) : ('a, 'b) rat_expr -> ('a, 'b) rat_expr -> ('a, 'b) bool_expr

val (>=.) : ('a, 'b) rat_expr -> ('a, 'b) rat_expr -> ('a, 'b) bool_expr

val (<=.) : ('a, 'b) rat_expr -> ('a, 'b) rat_expr -> ('a, 'b) bool_expr

val (==.) : ('a, 'b) rat_expr -> ('a, 'b) rat_expr -> ('a, 'b) bool_expr

val (!=.) : ('a, 'b) rat_expr -> ('a, 'b) rat_expr -> ('a, 'b) bool_expr

val rconst : Number.Rational.t -> ('a, 'b) rat_expr

val (+.) : ('a, 'b) rat_expr -> ('a, 'b) rat_expr -> ('a, 'b) rat_expr

val (-.) : ('a, 'b) rat_expr -> ('a, 'b) rat_expr -> ('a, 'b) rat_expr

val (*.) : ('a, 'b) rat_expr -> ('a, 'b) rat_expr -> ('a, 'b) rat_expr

val (/.) : ('a, 'b) rat_expr -> ('a, 'b) rat_expr -> ('a, 'b) rat_expr

val rite : ('a, 'b) bool_expr -> ('a, 'b) rat_expr -> ('a, 'b) rat_expr -> ('a, 'b) rat_expr

val r0 : ('a, 'b) rat_expr

Booleans

val t : ('a, 'b) bool_expr

True constant.

val f : ('a, 'b) bool_expr

False constant.

val binvar : 'a -> ('b, 'a) bool_expr

Boolean input variable.

val bsvar : 'a -> ('a, 'b) bool_expr

Boolean state variable.

val (&&) : ('a, 'b) bool_expr -> ('a, 'b) bool_expr -> ('a, 'b) bool_expr

Boolean conjunction.

val (||) : ('a, 'b) bool_expr -> ('a, 'b) bool_expr -> ('a, 'b) bool_expr

Boolean disjunction.

val (==>) : ('a, 'b) bool_expr -> ('a, 'b) bool_expr -> ('a, 'b) bool_expr

Boolean implication. x ==> y = !x || y

val (<=>) : ('a, 'b) bool_expr -> ('a, 'b) bool_expr -> ('a, 'b) bool_expr

Boolean equality (iff).

val (<!>) : ('a, 'b) bool_expr -> ('a, 'b) bool_expr -> ('a, 'b) bool_expr

Boolean inequality. a <!> b = !(a <=> y)

val (!) : ('a, 'b) bool_expr -> ('a, 'b) bool_expr

Boolean negation.

val bite : ('a, 'b) bool_expr -> ('a, 'b) bool_expr -> ('a, 'b) bool_expr -> ('a, 'b) bool_expr

Boolean if-then-else.

Integer queues

val iqueue : 'a -> ('a, 'b) int_queue_expr

val ipush : ('a, 'b) int_queue_expr -> ('a, 'b) int_expr -> ('a, 'b) int_queue_expr

val ipop : ('a, 'b) int_queue_expr -> ('a, 'b) int_queue_expr

val ifirst : 'a -> ('a, 'b) int_expr

val ilast : 'a -> ('a, 'b) int_expr

val all_positive : 'a -> ('a, 'b) bool_expr

val decrease : ('a, 'b) int_queue_expr -> ('a, 'b) int_queue_expr

val increase : ('a, 'b) int_queue_expr -> ('a, 'b) int_queue_expr

val iqite : ('a, 'b) bool_expr -> ('a, 'b) int_queue_expr -> ('a, 'b) int_queue_expr -> ('a, 'b) int_queue_expr

rational queues

val rqueue : 'a -> ('a, 'b) rat_queue_expr

val rpush : ('a, 'b) rat_queue_expr -> ('a, 'b) rat_expr -> ('a, 'b) rat_queue_expr

val rpop : ('a, 'b) rat_queue_expr -> ('a, 'b) rat_queue_expr

val rfirst : 'a -> ('a, 'b) rat_expr

val rlast : 'a -> ('a, 'b) rat_expr

val rqite : ('a, 'b) bool_expr -> ('a, 'b) rat_queue_expr -> ('a, 'b) rat_queue_expr -> ('a, 'b) rat_queue_expr

Machine description.

val (|->) : (('a, empty) bool_expr * ('a, 'b) bool_expr) -> ('a, 'b) assignment list -> ('a, 'b) transition

State and input guard maps into variable assignments.

val (@@@) : 'a -> 'b -> 'a * 'b

State guard and input guard.

val (&&&) : ('a, 'b) transition list -> ('a, empty) bool_expr -> ('a, 'b) t

Transition description with invariant.

val (=) : 'a -> ('b, 'c) int_expr -> 'a * ('b, 'c) expr

Assigns integer expression to an integer variable.

val (=.) : 'a -> ('b, 'c) rat_expr -> 'a * ('b, 'c) expr

Assigns rational expression to a rational variable.

val (=&) : 'a -> ('b, 'c) bool_expr -> 'a * ('b, 'c) expr

Assigns Boolean expression to a Boolean variable.

val (=|) : 'a -> ('b, 'c) int_queue_expr -> 'a * ('b, 'c) expr

Assigns integer queue expression to an integer queue variable.

val (=|.) : 'a -> ('b, 'c) rat_queue_expr -> 'a * ('b, 'c) expr

Assigns rational queue expression to an rational queue variable.