Module CausalEffect.EventEqTransition

Parameters

module T : sig ... end

module Time : Time

Signature

include sig ... end

module Place : sig ... end

Place IDs.

module Transition : sig ... end

Transition IDs.

module Color : sig ... end

Color IDs.

module Event : sig ... end

Event IDs.

module Probe : sig ... end

Probe IDs.

module Places : sig ... end

module Transitions : sig ... end

module Events : sig ... end

module Probes : sig ... end

type place_type = Network(T)(Time).place_type =

| Variable

| Queue

module Coloring : sig ... end

type token = Network(T)(Time).token = {

visits : Time.t Prelude.Dynarray.t Events.t;

coloring : Coloring.t;

}

Token type. Represents data passing through a system. Witnesses the relationships between events from the data interactions.

val compare_token : token -> token -> int

val token_of_sexp : Sexplib0.Sexp.t -> token

val sexp_of_token : token -> Sexplib0.Sexp.t

val empty_token : token

val merge_tokens : token -> token -> token

val add_visit : token -> Events.key -> Time.t -> Coloring.t -> token

type inner_place = Network(T)(Time).inner_place =

| Variable of token * bool

| Queue of token list

Variable type.

val read_from_place : inner_place -> token option * inner_place

val write_to_place : token -> inner_place -> inner_place

type inner_transition = Network(T)(Time).inner_transition = {

input_arcs : Place.t list;

output_arcs : Place.t list;

coloring : Coloring.t;

event : Event.t;

}

type color_match = Network(T)(Time).color_match =

| Any

| Single of Color.t

| And of color_match * color_match

| Or of color_match * color_match

Type of color matches.

val compare_color_match : color_match -> color_match -> int

val eval_color_match : Stdlib__Set.Make(Color).t -> color_match -> bool

type t = Network(T)(Time).t = {

places : inner_place Places.t;

transitions : inner_transition Transitions.t;

probes : (color_match * Transition.t) Probes.t;

}

Network type.

val empty : t

exception IncompatiblePlaceDefinition of Place.t

Raised when code declares that a place is both variable and queue at the same time.

exception MissingTransition of Transition.t

Raised when transition is expected, but not found.

exception IncompatibleTransitionDefinition of Transition.t

Raised when transition is declared twice with different events.

exception IncompatibleProbeDefinition of Probe.t

Raised when probe is declared twice with different matching colors or at different transition.

val add_variable : Places.key -> t -> t

val add_queue : Places.key -> t -> t

val add_transition : Transitions.key -> Event.t -> t -> t

val modify_transition : Transitions.key -> (inner_transition -> inner_transition) -> t -> t

val add_source : Transitions.key -> Place.t -> t -> t

val add_target : Transitions.key -> Place.t -> t -> t

val inject_color : Transitions.key -> Coloring.elt -> t -> t

val add_probe : Probes.key -> color_match -> Transition.t -> t -> t

module Compiled : sig ... end

Module for networks ready to travers traces efficiently. The representation is mutable.

val declare_data_connection : (Place.t -> 'a -> t) -> Place.t -> Transitions.key list -> Transitions.key list -> 'a -> t

Adds a place under a variable and arcs in transitions reading and writing to the place.

val sampling_variable : Place.t -> Transitions.key list -> Transitions.key list -> t -> t

Established an order-dependent one-to-many relation between write and read events.

val queueing : Place.t -> Transitions.key list -> Transitions.key list -> t -> t

Establishes a one-to-one relationship between the write and read events.

module Declaration : sig ... end

val witness_interval : token -> Time.t option

Largest difference between the witnessed event timestampts.

val shortest_longest_chains : Events.key -> token Prelude.Dynarray.t -> ((token * Time.t option) * (token * Time.t option)) Prelude.Seq.t

Computes shortest and longest event chains that pass through the same event.

exception MissingEvent of Event.t

val full_interval : Event.t -> (unit -> (token * 'a) Prelude.Seq.node) -> ((token * 'a) * Time.t) Prelude.Seq.t

Adds full interval causal link. Corresponds to the time between the event and its previous appearence with a successful chain.

exception TraceShorterThanChain

Impossible situation, trace finished before chain.

val reduced_interval : to_seq:('a -> Event.t Prelude.Seq.t) -> Event.t -> token Prelude.Seq.t -> ('a, Time.t) Trace.step Prelude.Seq.t -> (token * Time.t) Prelude.Seq.t

Adds reduced interval causal link. Corresponds to the time between the event and its previous appearence.